+++ aim/web/ui/src/components/HighPlot/HighPlot.tsx
@@ -89,6 +89,7 @@ const HighPlot = React.forwardRef(function HighPlot(
       dimensions: data.dimensions,
       plotBoxRef,
     });
+
     if (attributesRef?.current.xScale && attributesRef.current.yScale) {
       drawParallelLines({
         linesNodeRef,