+++ aim/web/ui/src/pages/RunDetail/RunDetail.tsx
@@ -197,6 +197,7 @@ function RunDetail(): React.FunctionComponentElement<React.ReactNode> {
             <Tab label='Images' />
             <Tab label='Distributions' />
             <Tab label='Texts' />
+            <Tab label='Plotly' />
             <Tab label='Settings' />
           </Tabs>
         </Paper>
@@ -272,6 +273,17 @@ function RunDetail(): React.FunctionComponentElement<React.ReactNode> {
         <TabPanel
           value={value}
           index={6}
+          className='RunDetail__runDetailContainer__tabPanel'
+        >
+          <TraceVisualizationContainer
+            runHash={runHash}
+            traceType='figures'
+            traceInfo={runData?.runTraces}
+          />
+        </TabPanel>
+        <TabPanel
+          value={value}
+          index={7}
           className='RunDetail__runDetailContainer__tabPanel'
         >
           <RunDetailSettingsTab