+++ aim/web/ui/src/pages/RunDetail/TraceVisualizationContainer/TraceVisualizationContainer.tsx
@@ -8,8 +8,9 @@ import useModel from 'hooks/model/useModel';
 import runTracesModel from 'services/models/runs/runTracesModel';
 
 import DistributionsVisualizer from '../DistributionsVisualizer';
-import ImagesVisualizer from '../ImagesVisualizer/ImagesVisualizer';
 import TextsVisualizer from '../TextsVisualizer';
+import ImagesVisualizer from '../ImagesVisualizer';
+import PlotlyVisualizer from '../PlotlyVisualizer';
 import { ITraceVisualizationContainerProps } from '../types';
 
 import RangePanel from './RangePanel';
@@ -23,7 +24,7 @@ const traceTypeVisualization = {
   audios: () => null,
   videos: () => null,
   texts: TextsVisualizer,
-  plotly: () => null,
+  figures: PlotlyVisualizer,
 };
 
 function TraceVisualizationContainer({
@@ -88,9 +89,10 @@ function TraceVisualizationContainer({
                 rangeEndpoints: runTracesModelData?.data[item.sliderName],
                 selectedRangeValue: runTracesModelData?.queryData?.sliders[
                   item.sliderName
-                ] || [0, 50],
+                ] || [0, 0],
                 inputValue:
-                  runTracesModelData?.queryData?.inputs[item.inputName] || 50,
+                  runTracesModelData?.queryData?.inputs[item.inputName] || 0,
+                sliderType: item.sliderType,
               }))}
               onApply={runTracesModel.onApply}
               onInputChange={runTracesModel.onInputChange}