+++ aim/web/ui/src/pages/RunDetail/types.d.ts
@@ -86,3 +86,5 @@ export interface IImagesVisualizerProps extends ITraceVisualizerProps {}
 export interface IDistributionVisualizerProps extends ITraceVisualizerProps {}
 
 export interface ITextsVisualizerProps extends ITraceVisualizerProps {}
+
+export interface IPlotlyVisualizerProps extends ITraceVisualizerProps {}