+++ aim/web/ui/src/services/models/runs/types.d.ts
@@ -28,7 +28,7 @@ export type TraceType =
   | 'audios'
   | 'texts'
   | 'videos'
-  | 'plotly';
+  | 'figures';
 
 /**
  * The context of info Raw data
@@ -83,6 +83,22 @@ export type DistributionValue = {
   };
 };
 
+/**
+ * Plotly api response
+ */
+export interface IPlotlyData extends TraceRawDataItem {
+  record_range: [number, number];
+  values: PlotlyValue[];
+  iters: number[];
+}
+
+/**
+ * Plotly api response value type
+ */
+export type IPlotlyValue = {
+  data: any[];
+  layout: any;
+};
 export interface TraceProcessedData extends DistributionValue {
   data: {
     blob: TraceProcessedValue;
@@ -98,14 +114,16 @@ export type QueryData = {
   inputs: Record<string, number>;
 };
 
-type RangePanelItem = {
+export type RangePanelItem = {
   sliderName: string;
   inputName: string;
   sliderTitle: string;
   inputTitle: string;
   sliderTitleTooltip: string;
   inputTitleTooltip: string;
+  sliderType: 'single' | 'range'; // This type is same as SliderWithInput component sliderType prop type.
 };
+
 export type IConfig = {
   rangePanel: RangePanelItem[];
 };