+++ aim/web/ui/src/services/models/runs/util.ts
@@ -1,5 +1,5 @@
 import bs58check from 'bs58check';
-import _ from 'lodash';
+import { head, orderBy, get } from 'lodash-es';
 
 import { IMenuItem } from 'components/kit/Menu';
 
@@ -18,6 +18,7 @@ import imagesExploreAppModel from '../imagesExplore/imagesExploreAppModel';
 import {
   DistributionsData,
   DistributionValue,
+  IPlotlyData,
   TraceProcessedData,
   TraceRawDataItem,
   TraceType,
@@ -185,7 +186,7 @@ export function getMenuData(traceType: TraceType, traces: TraceRawDataItem[]) {
     audios: 'Audios',
     videos: 'Videos',
     texts: 'Texts',
-    plotly: 'Plotlies',
+    figures: 'Plotlies',
   };
 
   let title = VisualizationMenuTitles[traceType];
@@ -312,7 +313,7 @@ export function processImagesData(
     });
   });
   const { imageSetData, orderedMap } = imagesExploreAppModel.getDataAsImageSet(
-    groupData(_.orderBy(images)),
+    groupData(orderBy(images)),
     groupingSelectOptions,
     ['step'],
   );
@@ -335,7 +336,7 @@ function groupData(data: IProcessedImageData[]): {
   for (let i = 0; i < data.length; i++) {
     const groupValue: { [key: string]: string } = {};
     ['step'].forEach((field) => {
-      groupValue[field] = _.get(data[i], field);
+      groupValue[field] = get(data[i], field);
     });
     const groupKey = encode(groupValue);
     if (groupValues.hasOwnProperty(groupKey)) {
@@ -361,4 +362,22 @@ export function reformatArrayQueries(
   });
 
   return formattedQueryObject;
+}
+
+/**
+ * process plotly data
+ */
+export function processPlotlyData(data: Partial<IPlotlyData>) {
+  const { record_range, iters, values } = data;
+  const processedValue = head(values);
+  const originalValues = values;
+
+  processedValue.layout.autosize = true;
+
+  return {
+    iters,
+    record_range,
+    processedValue,
+    originalValues,
+  };
 }