+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add Scatter Explorer to visualize correlations between metric last value and hyperparameter (KaroMourad)
+- Add ability to track and visualize plotly objects (devfox-se, Hamik25, rubenaprikyan)
 - Add ability to query distributions by step range and density (VkoHov, rubenaprikyan)
 - Implement images visualization tab in run detail page (VkoHov, KaroMourad)